refactor(web): split 5 large components into folder-based modules
Break down oversized single-file components into focused sub-components
following the existing ContractCreatePage/StockPage folder pattern:

- ContractDetailPage (1015 lines) → index + ContractInfoSection,
  PaymentScheduleTab, SignaturesSection, useContractDetail, useContractActions
- PaymentsPage (905 lines) → index + PendingPaymentsTab, PaymentSummaryTab,
  RecordPaymentModal, SlipScanner, usePaymentOcr hook (deduplicates 3x OCR)
- CustomersPage (917 lines) → index + CustomerTable, CustomerFormModal,
  CustomerFilters
- DashboardPage (814 lines) → index + KPICards, MonthlyTrendChart,
  TopOverdueTable, BranchComparisonTable
- Sidebar (550 lines) → index + SidebarMenuItems, SidebarNavGroup,
  SidebarUserInfo

Also extracts shared QueryErrorBlock component from DashboardPage.
