+++ apps/web/src/components/layout/Sidebar/index.tsx
@@ -1,0 +1,320 @@
+import { useMemo, useCallback, memo, useState } from 'react';
+import { Link, useLocation } from 'react-router-dom';
+import { useAuth } from '@/contexts/AuthContext';
+import { cn } from '@/lib/utils';
+import {
+  Home,
+  LogOut,
+  ChevronsRight,
+  ChevronsLeft,
+} from 'lucide-react';
+import {
+  Tooltip,
+  TooltipContent,
+  TooltipProvider,
+  TooltipTrigger,
+} from '@/components/ui/tooltip';
+import {
+  Popover,
+  PopoverContent,
+  PopoverTrigger,
+} from '@/components/ui/popover';
+import { useLayout } from '../LayoutContext';
+import { navSections } from './SidebarMenuItems';
+import type { NavSection } from './SidebarMenuItems';
+import { SidebarNavGroup } from './SidebarNavGroup';
+import { SidebarUserInfo } from './SidebarUserInfo';
+
+/* ─── Collapsed Icon Rail (70px) — Demo 9 default ─── */
+function CollapsedSidebar({ onToggle }: { onToggle: () => void }) {
+  const { user, logout } = useAuth();
+  const { pathname } = useLocation();
+  const [openPopover, setOpenPopover] = useState<string | null>(null);
+
+  const filteredSections = useMemo((): NavSection[] => {
+    return navSections
+      .map((section) => ({
+        ...section,
+        items: section.items.filter(
+          (item) => !item.roles || (user && item.roles.includes(user.role)),
+        ),
+      }))
+      .filter((section) => section.items.length > 0);
+  }, [user]);
+
+  const isSectionActive = useCallback(
+    (section: NavSection): boolean => {
+      return section.items.some(
+        (item) => item.path === pathname || (item.path.length > 1 && (pathname.startsWith(item.path + '/') || pathname === item.path)),
+      );
+    },
+    [pathname],
+  );
+
+  const isItemActive = useCallback(
+    (path: string): boolean =>
+      path === pathname || (path.length > 1 && (pathname.startsWith(path + '/') || pathname === path)),
+    [pathname],
+  );
+
+  return (
+    <div className="sidebar fixed top-0 bottom-0 left-0 z-20 w-[70px] flex flex-col items-center bg-sidebar-dark py-5 gap-1 transition-all duration-300">
+      {/* Logo */}
+      <Link to="/" className="flex items-center justify-center mb-2 shrink-0">
+        <div className="size-10 rounded-xl bg-primary flex items-center justify-center">
+          <span className="text-white text-lg font-bold">B</span>
+        </div>
+      </Link>
+
+      {/* Expand toggle */}
+      <button
+        onClick={onToggle}
+        className="flex items-center justify-center size-8 rounded-lg text-white/30 hover:text-white hover:bg-white/10 transition-colors mb-2"
+        aria-label="ขยายเมนู"
+      >
+        <ChevronsRight className="size-4" />
+      </button>
+
+      {/* Home icon */}
+      <TooltipProvider delayDuration={0}>
+        <Tooltip>
+          <TooltipTrigger asChild>
+            <Link
+              to="/"
+              className={cn(
+                'flex items-center justify-center size-10 rounded-xl transition-colors',
+                pathname === '/'
+                  ? 'bg-primary text-white'
+                  : 'text-white/50 hover:text-white hover:bg-white/10',
+              )}
+            >
+              <Home className="size-5" />
+            </Link>
+          </TooltipTrigger>
+          <TooltipContent side="right" sideOffset={8}>
+            หน้าหลัก
+          </TooltipContent>
+        </Tooltip>
+      </TooltipProvider>
+
+      {/* Section icons with popovers */}
+      <div className="flex-1 flex flex-col items-center gap-1 py-2 overflow-y-auto">
+        <TooltipProvider delayDuration={0}>
+          {filteredSections.map((section) => (
+            <Popover
+              key={section.key}
+              open={openPopover === section.key}
+              onOpenChange={(open) => setOpenPopover(open ? section.key : null)}
+            >
+              <Tooltip>
+                <TooltipTrigger asChild>
+                  <PopoverTrigger asChild>
+                    <button
+                      className={cn(
+                        'flex items-center justify-center size-10 rounded-xl transition-colors',
+                        isSectionActive(section)
+                          ? 'bg-primary text-white'
+                          : 'text-white/50 hover:text-white hover:bg-white/10',
+                      )}
+                    >
+                      <section.icon className="size-5" />
+                    </button>
+                  </PopoverTrigger>
+                </TooltipTrigger>
+                {openPopover !== section.key && (
+                  <TooltipContent side="right" sideOffset={8}>
+                    {section.label}
+                  </TooltipContent>
+                )}
+              </Tooltip>
+              <PopoverContent
+                side="right"
+                sideOffset={12}
+                align="start"
+                className="w-56 p-2"
+              >
+                <div className="text-xs font-semibold text-muted-foreground uppercase tracking-wider px-2 py-1.5 mb-1">
+                  {section.label}
+                </div>
+                {section.items.map((item) => (
+                  <Link
+                    key={item.path}
+                    to={item.path}
+                    onClick={() => setOpenPopover(null)}
+                    className={cn(
+                      'flex items-center gap-2.5 px-2 py-2 rounded-lg text-sm transition-colors',
+                      isItemActive(item.path)
+                        ? 'bg-primary/10 text-primary font-medium'
+                        : 'text-foreground/80 hover:bg-muted hover:text-foreground',
+                    )}
+                  >
+                    {item.icon && <item.icon className="size-4 shrink-0" />}
+                    <span>{item.label}</span>
+                  </Link>
+                ))}
+              </PopoverContent>
+            </Popover>
+          ))}
+        </TooltipProvider>
+      </div>
+
+      {/* Bottom: User avatar + logout */}
+      <div className="flex flex-col items-center gap-2 mt-auto pt-2">
+        {user && (
+          <TooltipProvider delayDuration={0}>
+            <Tooltip>
+              <TooltipTrigger asChild>
+                <div className="size-9 rounded-full bg-primary/30 flex items-center justify-center cursor-default">
+                  <span className="text-white text-sm font-semibold">{user.name?.charAt(0)}</span>
+                </div>
+              </TooltipTrigger>
+              <TooltipContent side="right" sideOffset={8}>
+                <div>
+                  <p className="font-medium">{user.name}</p>
+                  <p className="text-xs text-muted-foreground">{user.branchName}</p>
+                </div>
+              </TooltipContent>
+            </Tooltip>
+          </TooltipProvider>
+        )}
+        <TooltipProvider delayDuration={0}>
+          <Tooltip>
+            <TooltipTrigger asChild>
+              <button
+                onClick={logout}
+                className="flex items-center justify-center size-10 rounded-xl text-white/40 hover:text-white hover:bg-white/10 transition-colors"
+              >
+                <LogOut className="size-5" />
+              </button>
+            </TooltipTrigger>
+            <TooltipContent side="right" sideOffset={8}>
+              ออกจากระบบ
+            </TooltipContent>
+          </Tooltip>
+        </TooltipProvider>
+      </div>
+    </div>
+  );
+}
+
+/* ─── Expanded Full Sidebar (264px) ─── */
+function ExpandedSidebar({ onToggle }: { onToggle: () => void }) {
+  const { user, logout } = useAuth();
+  const { pathname } = useLocation();
+
+  const matchPath = useCallback(
+    (path: string): boolean =>
+      path === pathname || (path.length > 1 && pathname.startsWith(path)),
+    [pathname],
+  );
+
+  const filteredSections = useMemo((): NavSection[] => {
+    return navSections
+      .map((section) => ({
+        ...section,
+        items: section.items.filter(
+          (item) => !item.roles || (user && item.roles.includes(user.role)),
+        ),
+      }))
+      .filter((section) => section.items.length > 0);
+  }, [user]);
+
+  return (
+    <div className="sidebar fixed top-0 bottom-0 left-0 z-20 w-[264px] flex flex-col bg-sidebar-dark transition-all duration-300">
+      {/* Header */}
+      <div className="flex items-center justify-between px-5 h-[70px] shrink-0 border-b border-white/10">
+        <Link to="/" className="flex items-center gap-2.5">
+          <div className="size-9 rounded-lg bg-primary flex items-center justify-center">
+            <span className="text-white text-base font-bold">B</span>
+          </div>
+          <span className="text-base font-bold text-white leading-tight tracking-tight">
+            BEST<span className="text-primary">CHOICE</span>
+          </span>
+        </Link>
+        <button
+          onClick={onToggle}
+          className="flex items-center justify-center size-8 rounded-lg text-white/30 hover:text-white hover:bg-white/10 transition-colors"
+          aria-label="ย่อเมนู"
+        >
+          <ChevronsLeft className="size-4" />
+        </button>
+      </div>
+
+      {/* Navigation */}
+      <SidebarNavGroup
+        sections={filteredSections}
+        matchPath={matchPath}
+        selectedValue={pathname}
+      />
+
+      {/* User footer */}
+      {user && <SidebarUserInfo user={user} onLogout={logout} />}
+    </div>
+  );
+}
+
+/* ─── Mobile Full Sidebar (Sheet) ─── */
+function MobileSidebar() {
+  const { user, logout } = useAuth();
+  const { pathname } = useLocation();
+
+  const matchPath = useCallback(
+    (path: string): boolean =>
+      path === pathname || (path.length > 1 && pathname.startsWith(path)),
+    [pathname],
+  );
+
+  const filteredSections = useMemo((): NavSection[] => {
+    return navSections
+      .map((section) => ({
+        ...section,
+        items: section.items.filter(
+          (item) => !item.roles || (user && item.roles.includes(user.role)),
+        ),
+      }))
+      .filter((section) => section.items.length > 0);
+  }, [user]);
+
+  return (
+    <div className="w-full h-full flex flex-col bg-sidebar-dark">
+      {/* Header */}
+      <div className="flex items-center gap-2.5 px-5 h-[70px] shrink-0 border-b border-white/10">
+        <div className="size-9 rounded-lg bg-primary flex items-center justify-center">
+          <span className="text-white text-base font-bold">B</span>
+        </div>
+        <span className="text-base font-bold text-white leading-tight tracking-tight">
+          BEST<span className="text-primary">CHOICE</span>
+        </span>
+      </div>
+
+      {/* Navigation */}
+      <SidebarNavGroup
+        sections={filteredSections}
+        matchPath={matchPath}
+        selectedValue={pathname}
+      />
+
+      {/* User footer */}
+      {user && <SidebarUserInfo user={user} onLogout={logout} />}
+    </div>
+  );
+}
+
+/* ─── Main Sidebar Component ─── */
+function Sidebar({ mobile = false }: { mobile?: boolean }) {
+  const { sidebarCollapse, setSidebarCollapse } = useLayout();
+
+  if (mobile) {
+    return <MobileSidebar />;
+  }
+
+  const handleToggle = () => setSidebarCollapse(!sidebarCollapse);
+
+  if (sidebarCollapse) {
+    return <CollapsedSidebar onToggle={handleToggle} />;
+  }
+
+  return <ExpandedSidebar onToggle={handleToggle} />;
+}
+
+export default memo(Sidebar);
