+++ apps/web/src/pages/ContractDetailPage/types.ts
@@ -1,0 +1,76 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+
+export interface Payment {
+  id: string;
+  installmentNo: number;
+  dueDate: string;
+  amountDue: string;
+  amountPaid: string | null;
+  lateFee: string;
+  status: string;
+  paidDate: string | null;
+  paymentMethod: string | null;
+}
+
+export interface ContractDetail {
+  id: string;
+  contractNumber: string;
+  status: string;
+  workflowStatus: string;
+  planType: string;
+  sellingPrice: string;
+  downPayment: string;
+  interestRate: string;
+  totalMonths: number;
+  interestTotal: string;
+  financedAmount: string;
+  monthlyPayment: string;
+  paymentDueDay: number | null;
+  notes: string | null;
+  reviewNotes: string | null;
+  contractHash: string | null;
+  pdpaConsentId: string | null;
+  createdAt: string;
+  reviewedAt: string | null;
+  salespersonId: string;
+  customer: { id: string; name: string; phone: string; nationalId: string };
+  customerSnapshot: { name: string; phone: string; nationalId?: string; prefix?: string; nickname?: string; occupation?: string; salary?: string } | null;
+  product: { id: string; name: string; brand: string; model: string; category: string; color: string | null; storage: string | null; serialNumber: string | null; imeiSerial: string | null; costPrice: string; batteryHealth: number | null; warrantyExpired: boolean | null; warrantyExpireDate: string | null; hasBox: boolean | null; accessoryType: string | null; accessoryBrand: string | null };
+  branch: { id: string; name: string };
+  salesperson: { id: string; name: string };
+  reviewedBy: { id: string; name: string } | null;
+  interestConfig: { id: string; name: string; storeCommissionPct?: string; vatPct?: string } | null;
+  creditBalance: string | null;
+  dunningStage: string | null;
+  payments: Payment[];
+  signatures: { id: string; signerType: string; signedAt: string }[];
+  contractDocuments: any[];
+  creditCheck: any;
+}
+
+export interface EarlyPayoffQuote {
+  remainingMonths: number;
+  remainingPrincipal: number;
+  remainingInterest: number;
+  discount: number;
+  unpaidLateFees: number;
+  totalPayoff: number;
+}
+
+export const statusLabels: Record<string, { label: string; className: string }> = {
+  DRAFT: { label: 'ร่าง', className: 'bg-secondary text-foreground' },
+  ACTIVE: { label: 'ผ่อนอยู่', className: 'bg-green-100 text-green-700' },
+  OVERDUE: { label: 'ค้างชำระ', className: 'bg-yellow-100 text-yellow-700' },
+  DEFAULT: { label: 'ผิดนัด', className: 'bg-red-100 text-red-700' },
+  EARLY_PAYOFF: { label: 'ปิดก่อน', className: 'bg-primary/10 text-primary' },
+  COMPLETED: { label: 'ครบ', className: 'bg-teal-100 text-teal-700' },
+  EXCHANGED: { label: 'เปลี่ยนเครื่อง', className: 'bg-primary/10 text-primary' },
+  CLOSED_BAD_DEBT: { label: 'หนี้สูญ', className: 'bg-red-200 text-red-800' },
+};
+
+export const paymentStatusLabels: Record<string, { label: string; className: string }> = {
+  PENDING: { label: 'รอชำระ', className: 'bg-secondary text-foreground' },
+  PAID: { label: 'ชำระแล้ว', className: 'bg-green-100 text-green-700' },
+  OVERDUE: { label: 'เกินกำหนด', className: 'bg-red-100 text-red-700' },
+  PARTIALLY_PAID: { label: 'ชำระบางส่วน', className: 'bg-yellow-100 text-yellow-700' },
+};
