+++ apps/web/src/pages/CustomersPage/types.ts
@@ -1,0 +1,66 @@
+export interface Customer {
+  id: string;
+  nationalId: string;
+  name: string;
+  nickname: string | null;
+  phone: string;
+  lineId: string | null;
+  occupation: string | null;
+  salary: number | null;
+  createdAt: string;
+  _count: { contracts: number };
+  activeContracts: number;
+  overdueContracts: number;
+  latestCreditStatus: string | null;
+  latestCreditScore: number | null;
+}
+
+export interface CustomerSummary {
+  totalCustomers: number;
+  withActiveContract: number;
+  withOverdue: number;
+  newThisMonth: number;
+}
+
+export interface CustomersResponse {
+  data: Customer[];
+  total: number;
+  page: number;
+  limit: number;
+  totalPages: number;
+  summary: CustomerSummary;
+}
+
+export interface ReferenceData {
+  prefix: string;
+  firstName: string;
+  lastName: string;
+  phone: string;
+  relationship: string;
+}
+
+export const emptyReference: ReferenceData = { prefix: '', firstName: '', lastName: '', phone: '', relationship: '' };
+
+export const emptyForm = {
+  prefix: '',
+  firstName: '',
+  lastName: '',
+  nickname: '',
+  nationalId: '',
+  isForeigner: false,
+  birthDate: '',
+  phone: '',
+  phoneSecondary: '',
+  email: '',
+  lineId: '',
+  facebookLink: '',
+  facebookName: '',
+  facebookFriends: '',
+  googleMapLink: '',
+  occupation: '',
+  occupationDetail: '',
+  salary: '',
+  workplace: '',
+};
+
+export type CustomerForm = typeof emptyForm;
