+++ apps/web/src/pages/DashboardPage/BranchComparisonTable.tsx
@@ -1,0 +1,78 @@
+import { Card, CardContent, CardHeader, CardTitle, CardToolbar, CardTable } from '@/components/ui/card';
+import QueryErrorBlock from '@/components/ui/QueryErrorBlock';
+
+interface BranchComparison {
+  id: string;
+  name: string;
+  contracts: number;
+  products: number;
+  users: number;
+  overdueContracts: number;
+  monthlyPayments: number;
+}
+
+interface BranchComparisonTableProps {
+  branchData: BranchComparison[];
+  branchError: boolean;
+  refetchBranch: () => void;
+}
+
+export default function BranchComparisonTable({ branchData, branchError, refetchBranch }: BranchComparisonTableProps) {
+  return (
+    <>
+      {branchError && (
+        <Card>
+          <CardContent>
+            <QueryErrorBlock message="โหลดข้อมูลสาขาไม่สำเร็จ" onRetry={() => refetchBranch()} />
+          </CardContent>
+        </Card>
+      )}
+      {branchData.length > 0 && (
+        <Card>
+          <CardHeader>
+            <CardTitle>เปรียบเทียบสาขา</CardTitle>
+            <CardToolbar>
+              <span className="text-2xs text-muted-foreground bg-muted px-2.5 py-1 rounded-md font-medium">
+                {branchData.length} สาขา
+              </span>
+            </CardToolbar>
+          </CardHeader>
+          <CardTable>
+            <div className="overflow-x-auto">
+              <table className="w-full text-sm">
+                <thead>
+                  <tr className="border-b border-border/50 text-left text-muted-foreground">
+                    <th className="px-5 pb-3 pt-4 font-medium text-xs">สาขา</th>
+                    <th className="px-5 pb-3 pt-4 font-medium text-xs text-right">สัญญา</th>
+                    <th className="px-5 pb-3 pt-4 font-medium text-xs text-right">สินค้า</th>
+                    <th className="px-5 pb-3 pt-4 font-medium text-xs text-right">พนักงาน</th>
+                    <th className="px-5 pb-3 pt-4 font-medium text-xs text-right">ค้างชำระ</th>
+                    <th className="px-5 pb-3 pt-4 font-medium text-xs text-right">ยอดชำระ/เดือน</th>
+                  </tr>
+                </thead>
+                <tbody>
+                  {branchData.map((b) => (
+                    <tr key={b.name} className="border-b border-border/30 last:border-0 hover:bg-muted/50 transition-colors">
+                      <td className="px-5 py-3 font-medium text-foreground">{b.name}</td>
+                      <td className="px-5 py-3 text-right text-foreground">{b.contracts}</td>
+                      <td className="px-5 py-3 text-right text-foreground">{b.products}</td>
+                      <td className="px-5 py-3 text-right text-foreground">{b.users}</td>
+                      <td className="px-5 py-3 text-right">
+                        <span className={b.overdueContracts > 0 ? 'text-destructive font-semibold' : 'text-foreground'}>
+                          {b.overdueContracts}
+                        </span>
+                      </td>
+                      <td className="px-5 py-3 text-right text-success font-medium">
+                        {b.monthlyPayments.toLocaleString()}
+                      </td>
+                    </tr>
+                  ))}
+                </tbody>
+              </table>
+            </div>
+          </CardTable>
+        </Card>
+      )}
+    </>
+  );
+}
