+++ apps/web/src/pages/DashboardPage/TopOverdueTable.tsx
@@ -1,0 +1,83 @@
+import { Card, CardContent, CardHeader, CardTitle, CardToolbar, CardTable } from '@/components/ui/card';
+import { cn } from '@/lib/utils';
+import QueryErrorBlock from '@/components/ui/QueryErrorBlock';
+
+interface TopOverdue {
+  contractNumber: string;
+  customer: { id: string; name: string; phone: string };
+  totalOutstanding: number;
+  daysOverdue: number;
+}
+
+interface TopOverdueTableProps {
+  topOverdue: TopOverdue[];
+  topOverdueError: boolean;
+  refetchTopOverdue: () => void;
+}
+
+export default function TopOverdueTable({ topOverdue, topOverdueError, refetchTopOverdue }: TopOverdueTableProps) {
+  return (
+    <>
+      {topOverdueError && (
+        <Card>
+          <CardContent>
+            <QueryErrorBlock message="โหลดข้อมูลค้างชำระไม่สำเร็จ" onRetry={() => refetchTopOverdue()} />
+          </CardContent>
+        </Card>
+      )}
+      {topOverdue.length > 0 && (
+        <Card>
+          <CardHeader>
+            <CardTitle>สัญญาค้างชำระสูงสุด (Top 10)</CardTitle>
+            <CardToolbar>
+              <span className="text-2xs text-destructive bg-destructive/10 px-2.5 py-1 rounded-md font-medium">
+                {topOverdue.length} รายการ
+              </span>
+            </CardToolbar>
+          </CardHeader>
+          <CardTable>
+            <div className="overflow-x-auto">
+              <table className="w-full text-sm">
+                <thead>
+                  <tr className="border-b border-border/50 text-left text-muted-foreground">
+                    <th className="px-5 pb-3 pt-4 font-medium text-xs">เลขสัญญา</th>
+                    <th className="px-5 pb-3 pt-4 font-medium text-xs">ลูกค้า</th>
+                    <th className="px-5 pb-3 pt-4 font-medium text-xs">เบอร์โทร</th>
+                    <th className="px-5 pb-3 pt-4 font-medium text-xs text-right">ยอดค้าง (บาท)</th>
+                    <th className="px-5 pb-3 pt-4 font-medium text-xs text-right">เกินกำหนด</th>
+                  </tr>
+                </thead>
+                <tbody>
+                  {topOverdue.map((item) => (
+                    <tr key={item.contractNumber} className="border-b border-border/30 last:border-0 hover:bg-muted/50 transition-colors">
+                      <td className="px-5 py-3 font-medium text-primary">{item.contractNumber}</td>
+                      <td className="px-5 py-3 text-foreground">{item.customer.name}</td>
+                      <td className="px-5 py-3 text-muted-foreground">{item.customer.phone}</td>
+                      <td className="px-5 py-3 text-right text-destructive font-semibold">
+                        {item.totalOutstanding.toLocaleString()}
+                      </td>
+                      <td className="px-5 py-3 text-right">
+                        <span
+                          className={cn(
+                            'px-2 py-0.5 rounded-md text-2xs font-medium',
+                            item.daysOverdue > 60
+                              ? 'bg-destructive/10 text-destructive'
+                              : item.daysOverdue > 30
+                                ? 'bg-warning/10 text-warning'
+                                : 'bg-orange-100 text-orange-700 dark:bg-orange-900/30 dark:text-orange-400',
+                          )}
+                        >
+                          {item.daysOverdue} วัน
+                        </span>
+                      </td>
+                    </tr>
+                  ))}
+                </tbody>
+              </table>
+            </div>
+          </CardTable>
+        </Card>
+      )}
+    </>
+  );
+}
