+++ apps/web/src/pages/DashboardPage/index.tsx
@@ -3,7 +3,7 @@ import { useAuth } from '@/contexts/AuthContext';
 import { useQuery } from '@tanstack/react-query';
 import { useNavigate } from 'react-router-dom';
 import api from '@/lib/api';
-import { Card, CardContent, CardHeader, CardTitle, CardToolbar, CardTable } from '@/components/ui/card';
+import { Card, CardContent, CardHeader, CardTitle, CardToolbar } from '@/components/ui/card';
 import { Tabs, TabsList, TabsTrigger, TabsContent } from '@/components/ui/tabs';
 import {
   ShoppingCart,
@@ -13,13 +13,17 @@ import {
   Warehouse,
   BarChart3,
   AlertTriangle,
-  TrendingUp,
   ArrowRight,
   Clock,
   UserCheck,
 } from 'lucide-react';
 import type { LucideIcon } from 'lucide-react';
 import { cn } from '@/lib/utils';
+import QueryErrorBlock from '@/components/ui/QueryErrorBlock';
+import KPICards from './KPICards';
+import MonthlyTrendChart from './MonthlyTrendChart';
+import TopOverdueTable from './TopOverdueTable';
+import BranchComparisonTable from './BranchComparisonTable';
 
 /* ─── Types ─── */
 
@@ -146,16 +150,6 @@ function ShortcutCard({ icon: Icon, label, path, color }: { icon: LucideIcon; la
   );
 }
 
-/* ─── Retry Error Block ─── */
-function ErrorBlock({ message, onRetry }: { message: string; onRetry: () => void }) {
-  return (
-    <div className="text-center py-8">
-      <p className="text-sm text-destructive mb-2">{message}</p>
-      <button onClick={onRetry} className="text-xs text-primary hover:underline">ลองใหม่</button>
-    </div>
-  );
-}
-
 /* ─── Time ago helper ─── */
 function timeAgo(dateStr: string): string {
   const diff = Date.now() - new Date(dateStr).getTime();
@@ -234,10 +228,6 @@ export default function DashboardPage() {
 
   /* ─── Computed ─── */
   const totalStatusCount = useMemo(() => statusDist.reduce((sum, s) => sum + s.count, 0), [statusDist]);
-  const trendMax = useMemo(() => {
-    if (trend.length === 0) return 1;
-    return Math.max(...trend.map((t) => Math.max(t.newContracts, t.paymentsReceived)), 1);
-  }, [trend]);
   const agingMax = useMemo(() => (aging ? Math.max(...aging.buckets.map((b) => b.amount), 1) : 1), [aging]);
 
   return (
@@ -260,47 +250,10 @@ export default function DashboardPage() {
         </div>
       )}
 
-      {/* ═══ KPI Banner (full-width) ═══ */}
-      {kpis && (
-        <div className="rounded-xl bg-gradient-to-br from-blue-500 via-blue-600 to-indigo-700 text-white p-6 lg:p-8">
-          <div className="grid grid-cols-2 lg:grid-cols-4 gap-6">
-            <div className="cursor-pointer" onClick={() => navigate('/contracts')}>
-              <div className="flex items-center gap-2 mb-2">
-                <FileCheck className="size-4 opacity-70" />
-                <span className="text-xs text-white/70 font-medium">สัญญาทั้งหมด</span>
-              </div>
-              <div className="text-2xl lg:text-3xl font-bold">{kpis.contracts.total}</div>
-              <div className="text-xs text-white/60 mt-1">ปกติ {kpis.contracts.active}</div>
-            </div>
-            <div className="cursor-pointer" onClick={() => navigate('/overdue')}>
-              <div className="flex items-center gap-2 mb-2">
-                <AlertTriangle className="size-4 opacity-70" />
-                <span className="text-xs text-white/70 font-medium">ค้าง/ผิดนัด</span>
-              </div>
-              <div className="text-2xl lg:text-3xl font-bold">{(kpis.contracts.overdue ?? 0) + (kpis.contracts.default ?? 0)}</div>
-              <div className="text-xs text-white/60 mt-1">{(kpis.overdueRate ?? 0).toFixed(1)}%</div>
-            </div>
-            <div className="cursor-pointer" onClick={() => navigate('/payments')}>
-              <div className="flex items-center gap-2 mb-2">
-                <TrendingUp className="size-4 opacity-70" />
-                <span className="text-xs text-white/70 font-medium">ยอดรับวันนี้</span>
-              </div>
-              <div className="text-2xl lg:text-3xl font-bold">฿{kpis.financial.todayPayments.toLocaleString()}</div>
-              <div className="text-xs text-white/60 mt-1">{kpis.financial.todayPaymentCount} รายการ</div>
-            </div>
-            <div className="cursor-pointer" onClick={() => navigate('/stock')}>
-              <div className="flex items-center gap-2 mb-2">
-                <Warehouse className="size-4 opacity-70" />
-                <span className="text-xs text-white/70 font-medium">สินค้าในสต็อก</span>
-              </div>
-              <div className="text-2xl lg:text-3xl font-bold">{kpis.products.inStock}</div>
-              <div className="text-xs text-white/60 mt-1">จาก {kpis.products.total}</div>
-            </div>
-          </div>
-        </div>
-      )}
+      {/* KPI Banner (full-width) */}
+      {kpis && <KPICards kpis={kpis} navigate={navigate} />}
 
-      {/* ═══ Two-Column: Shortcuts + Monthly Revenue ═══ */}
+      {/* Two-Column: Shortcuts + Monthly Revenue */}
       <div className="grid grid-cols-1 lg:grid-cols-12 gap-5 lg:gap-7">
         {/* Quick Action Shortcuts — role-based */}
         <div className="lg:col-span-5">
@@ -337,7 +290,7 @@ export default function DashboardPage() {
               </CardHeader>
               <CardContent className="p-0">
                 {revenueError ? (
-                  <ErrorBlock message="โหลดข้อมูลรายได้ไม่สำเร็จ" onRetry={() => refetchRevenue()} />
+                  <QueryErrorBlock message="โหลดข้อมูลรายได้ไม่สำเร็จ" onRetry={() => refetchRevenue()} />
                 ) : revenue ? (
                   <div className="divide-y divide-border/50">
                     <div className="flex items-center gap-4 px-5 py-3.5">
@@ -435,7 +388,7 @@ export default function DashboardPage() {
         </div>
       </div>
 
-      {/* ═══ Aging Buckets (full-width) — hide for SALES ═══ */}
+      {/* Aging Buckets (full-width) — hide for SALES */}
       {user?.role !== 'SALES' && <Card>
         <CardHeader>
           <CardTitle>อายุหนี้ค้างชำระ (Aging)</CardTitle>
@@ -449,7 +402,7 @@ export default function DashboardPage() {
         </CardHeader>
         <CardContent>
           {agingError ? (
-            <ErrorBlock message="โหลดข้อมูลอายุหนี้ไม่สำเร็จ" onRetry={() => refetchAging()} />
+            <QueryErrorBlock message="โหลดข้อมูลอายุหนี้ไม่สำเร็จ" onRetry={() => refetchAging()} />
           ) : aging ? (
             <div className="space-y-4">
               {aging.buckets.map((bucket) => (
@@ -477,7 +430,7 @@ export default function DashboardPage() {
         </CardContent>
       </Card>}
 
-      {/* ═══ Staff Performance (Tabs) — OWNER only ═══ */}
+      {/* Staff Performance (Tabs) — OWNER only */}
       {user?.role === 'OWNER' && <Card>
         <CardHeader>
           <CardTitle>กำกับพนักงาน</CardTitle>
@@ -491,7 +444,7 @@ export default function DashboardPage() {
         </CardHeader>
         <CardContent>
           {staffError ? (
-            <ErrorBlock message="โหลดข้อมูลพนักงานไม่สำเร็จ" onRetry={() => refetchStaff()} />
+            <QueryErrorBlock message="โหลดข้อมูลพนักงานไม่สำเร็จ" onRetry={() => refetchStaff()} />
           ) : staffPerf ? (
             <Tabs defaultValue="sales">
               <TabsList variant="line" size="sm">
@@ -597,57 +550,11 @@ export default function DashboardPage() {
         </CardContent>
       </Card>}
 
-      {/* ═══ Two-Column: Trend + Status Distribution ═══ */}
+      {/* Two-Column: Trend + Status Distribution */}
       <div className="grid grid-cols-1 lg:grid-cols-12 gap-5 lg:gap-7">
         {/* Monthly Trend */}
         <div className="lg:col-span-5">
-          <Card>
-            <CardHeader>
-              <CardTitle>แนวโน้ม 12 เดือน</CardTitle>
-              <CardToolbar>
-                <span className="text-2xs text-muted-foreground">Latest trends</span>
-              </CardToolbar>
-            </CardHeader>
-            <CardContent>
-              {trendError ? (
-                <ErrorBlock message="โหลดข้อมูลไม่สำเร็จ" onRetry={() => refetchTrend()} />
-              ) : trend.length > 0 ? (
-                <div className="space-y-2">
-                  {trend.map((t) => (
-                    <div key={t.month} className="flex items-center gap-3 text-xs">
-                      <div className="w-14 text-muted-foreground shrink-0 font-medium">{t.month}</div>
-                      <div className="flex-1 space-y-1">
-                        <div className="flex items-center gap-2">
-                          <div
-                            className="h-2.5 bg-primary rounded-full"
-                            style={{ width: `${(t.newContracts / trendMax) * 100}%`, minWidth: '2px' }}
-                          />
-                          <span className="text-foreground font-medium">{t.newContracts}</span>
-                        </div>
-                        <div className="flex items-center gap-2">
-                          <div
-                            className="h-2.5 bg-success rounded-full"
-                            style={{ width: `${(t.paymentsReceived / trendMax) * 100}%`, minWidth: '2px' }}
-                          />
-                          <span className="text-foreground font-medium">{t.paymentsReceived.toLocaleString()}</span>
-                        </div>
-                      </div>
-                    </div>
-                  ))}
-                  <div className="flex gap-4 text-2xs text-muted-foreground mt-4 pt-3 border-t border-border/50">
-                    <span className="flex items-center gap-1.5">
-                      <span className="size-2.5 bg-primary rounded-full inline-block" /> สัญญาใหม่
-                    </span>
-                    <span className="flex items-center gap-1.5">
-                      <span className="size-2.5 bg-success rounded-full inline-block" /> ยอดชำระ (บาท)
-                    </span>
-                  </div>
-                </div>
-              ) : (
-                <div className="text-center text-muted-foreground py-8 text-sm">ไม่มีข้อมูล</div>
-              )}
-            </CardContent>
-          </Card>
+          <MonthlyTrendChart trend={trend} trendError={trendError} refetchTrend={() => refetchTrend()} />
         </div>
 
         {/* Status Distribution */}
@@ -663,7 +570,7 @@ export default function DashboardPage() {
             </CardHeader>
             <CardContent>
               {statusDistError ? (
-                <ErrorBlock message="โหลดข้อมูลไม่สำเร็จ" onRetry={() => refetchStatusDist()} />
+                <QueryErrorBlock message="โหลดข้อมูลไม่สำเร็จ" onRetry={() => refetchStatusDist()} />
               ) : statusDist.length > 0 ? (
                 <div className="space-y-3">
                   {statusDist.map((s) => (
@@ -693,121 +600,12 @@ export default function DashboardPage() {
         </div>
       </div>
 
-      {/* ═══ Top Overdue Table ═══ */}
-      {topOverdueError && (
-        <Card>
-          <CardContent>
-            <ErrorBlock message="โหลดข้อมูลค้างชำระไม่สำเร็จ" onRetry={() => refetchTopOverdue()} />
-          </CardContent>
-        </Card>
-      )}
-      {topOverdue.length > 0 && (
-        <Card>
-          <CardHeader>
-            <CardTitle>สัญญาค้างชำระสูงสุด (Top 10)</CardTitle>
-            <CardToolbar>
-              <span className="text-2xs text-destructive bg-destructive/10 px-2.5 py-1 rounded-md font-medium">
-                {topOverdue.length} รายการ
-              </span>
-            </CardToolbar>
-          </CardHeader>
-          <CardTable>
-            <div className="overflow-x-auto">
-              <table className="w-full text-sm">
-                <thead>
-                  <tr className="border-b border-border/50 text-left text-muted-foreground">
-                    <th className="px-5 pb-3 pt-4 font-medium text-xs">เลขสัญญา</th>
-                    <th className="px-5 pb-3 pt-4 font-medium text-xs">ลูกค้า</th>
-                    <th className="px-5 pb-3 pt-4 font-medium text-xs">เบอร์โทร</th>
-                    <th className="px-5 pb-3 pt-4 font-medium text-xs text-right">ยอดค้าง (บาท)</th>
-                    <th className="px-5 pb-3 pt-4 font-medium text-xs text-right">เกินกำหนด</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {topOverdue.map((item) => (
-                    <tr key={item.contractNumber} className="border-b border-border/30 last:border-0 hover:bg-muted/50 transition-colors">
-                      <td className="px-5 py-3 font-medium text-primary">{item.contractNumber}</td>
-                      <td className="px-5 py-3 text-foreground">{item.customer.name}</td>
-                      <td className="px-5 py-3 text-muted-foreground">{item.customer.phone}</td>
-                      <td className="px-5 py-3 text-right text-destructive font-semibold">
-                        {item.totalOutstanding.toLocaleString()}
-                      </td>
-                      <td className="px-5 py-3 text-right">
-                        <span
-                          className={cn(
-                            'px-2 py-0.5 rounded-md text-2xs font-medium',
-                            item.daysOverdue > 60
-                              ? 'bg-destructive/10 text-destructive'
-                              : item.daysOverdue > 30
-                                ? 'bg-warning/10 text-warning'
-                                : 'bg-orange-100 text-orange-700 dark:bg-orange-900/30 dark:text-orange-400',
-                          )}
-                        >
-                          {item.daysOverdue} วัน
-                        </span>
-                      </td>
-                    </tr>
-                  ))}
-                </tbody>
-              </table>
-            </div>
-          </CardTable>
-        </Card>
-      )}
+      {/* Top Overdue Table */}
+      <TopOverdueTable topOverdue={topOverdue} topOverdueError={topOverdueError} refetchTopOverdue={() => refetchTopOverdue()} />
 
-      {/* ═══ Branch Comparison (OWNER only) ═══ */}
-      {user?.role === 'OWNER' && branchError && (
-        <Card>
-          <CardContent>
-            <ErrorBlock message="โหลดข้อมูลสาขาไม่สำเร็จ" onRetry={() => refetchBranch()} />
-          </CardContent>
-        </Card>
-      )}
-      {user?.role === 'OWNER' && branchData.length > 0 && (
-        <Card>
-          <CardHeader>
-            <CardTitle>เปรียบเทียบสาขา</CardTitle>
-            <CardToolbar>
-              <span className="text-2xs text-muted-foreground bg-muted px-2.5 py-1 rounded-md font-medium">
-                {branchData.length} สาขา
-              </span>
-            </CardToolbar>
-          </CardHeader>
-          <CardTable>
-            <div className="overflow-x-auto">
-              <table className="w-full text-sm">
-                <thead>
-                  <tr className="border-b border-border/50 text-left text-muted-foreground">
-                    <th className="px-5 pb-3 pt-4 font-medium text-xs">สาขา</th>
-                    <th className="px-5 pb-3 pt-4 font-medium text-xs text-right">สัญญา</th>
-                    <th className="px-5 pb-3 pt-4 font-medium text-xs text-right">สินค้า</th>
-                    <th className="px-5 pb-3 pt-4 font-medium text-xs text-right">พนักงาน</th>
-                    <th className="px-5 pb-3 pt-4 font-medium text-xs text-right">ค้างชำระ</th>
-                    <th className="px-5 pb-3 pt-4 font-medium text-xs text-right">ยอดชำระ/เดือน</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {branchData.map((b) => (
-                    <tr key={b.name} className="border-b border-border/30 last:border-0 hover:bg-muted/50 transition-colors">
-                      <td className="px-5 py-3 font-medium text-foreground">{b.name}</td>
-                      <td className="px-5 py-3 text-right text-foreground">{b.contracts}</td>
-                      <td className="px-5 py-3 text-right text-foreground">{b.products}</td>
-                      <td className="px-5 py-3 text-right text-foreground">{b.users}</td>
-                      <td className="px-5 py-3 text-right">
-                        <span className={b.overdueContracts > 0 ? 'text-destructive font-semibold' : 'text-foreground'}>
-                          {b.overdueContracts}
-                        </span>
-                      </td>
-                      <td className="px-5 py-3 text-right text-success font-medium">
-                        {b.monthlyPayments.toLocaleString()}
-                      </td>
-                    </tr>
-                  ))}
-                </tbody>
-              </table>
-            </div>
-          </CardTable>
-        </Card>
+      {/* Branch Comparison (OWNER only) */}
+      {user?.role === 'OWNER' && (
+        <BranchComparisonTable branchData={branchData} branchError={branchError} refetchBranch={() => refetchBranch()} />
       )}
     </div>
   );
